feat: move cytoscape config to a separate file
